Add 24hr support and fix setting manual alarms.
Closes #19
The file pylarm.py has been edited with a much
simpler code which is easier to read and
functions much more simply so that now the alarm
works in 24Hrs format.If the user sets alarm which is before
his local time, then the alarm automaatically sets the alarm
for the next day.If the user sets alarm for a time which is yet to
come on that day,i.e. after the local time,then the alarm sets itself
for that day only.
If the user isn't satisfied with the final time, he can manually set it.
The file manually_set_alarm.py has been edited and debugged
to add new parameters to manually_set_alarm function